requirements.txt
APScheduler==3.6.3
black==21.9b0
cachetools==4.2.2
certifi==2021.5.30
click==8.0.1
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
python-telegram-bot==13.7
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
six==1.16.0
toml==0.10.2
tomli==1.2.1
tornado==6.1
types-PyYAML==5.4.10
typing-extensions==3.10.0.2
tzlocal==3.0
